+++ NEWS.md
@@ -1,3 +1,8 @@
+# statswalesr 0.3.0
+
+* All functions deprecated except statswales_get_dataset() due to end
+of OData service
+
 # statswalesr 0.2.0
 
 * Welsh language options are now available across the three statswalesr functions